backend/requirements.txt
# Backend Dependencies - Python 3.13 compatible
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis>=5.0.1
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
aiosmtplib>=3.0.1
email-validator>=2.2.0
httpx>=0.27.0
